+++ styleguide.config.js
@@ -32,11 +32,7 @@ module.exports = {
 
   components: 'src/components/**/[A-Z]*.js',
 
-  require: [
-    path.join(__dirname, 'src/support-ie11.js'),
-    path.join(__dirname, 'src/globals/styles.js'),
-    path.join(__dirname, 'src/globals/chart-styles.js'),
-  ],
+  require: [path.join(__dirname, 'src/support-ie11.js')],
 
   // TODO: Get the right import statement: import { Component } from '@fashiontrade/wardrobe`.
   // Current Problem: DropdownItem should be used as Dropdown.Item but is imported from Dropdown
@@ -46,6 +42,7 @@ module.exports = {
 
   styleguideComponents: {
     Wrapper: path.join(__dirname, 'tools/ThemeWrapper'),
+    StyleGuideRenderer: path.join(__dirname, 'tools/StyleGuideRendererWrapper'),
   },
 
   styles: {